WIP steel threads example (#204)
* WIP steel threads example

* wip + fix markdown

* WIP

* Revise migration strategy and introduce Steel Threads

Updated the description of the migration strategy for The New York Times marketing system. Added details about the first Steel Thread and its implementation.

* Refine case study on NYT paywall modernization

Revised text for clarity and consistency in the case study on modernizing The New York Times paywall. Adjusted phrasing and structure for improved readability.

* Add files via upload

* Add image for Thread One in Steel Threads documentation

Added an image to enhance the content of the Steel Threads document.

* Add files via upload

* Stop Old Preview Versions

* steel threads

* template